Major changes
This commit includes a lot of changes, including:
Major UI/UX change, base template is changed
Most of pages have been reworked
Now it is possible to create new stores and wallets
without loading pages, just via popup.
Data is now represented in good-looking datatables(datatables.js)
which support paging, searching by any of columns of data,
and some other features.
In wallets page, simply press on wallet in the table on the left
to see transactions of this wallet in the table on the right
Another big change is support for two factor authentification.
Note that after some changes in our login templates it may not work
properly, submit an issue if it doesn't work. It is possible to
authentificate using any authentificator app like google
authentificator, or using backup codes generated.
Now we have changed to postgres database for better performance.
We have changed our structure, if before it was:
user->store->product
So now it is:
user->wallet->store->product
May look strange that you will need an extra step, but this provides
more flexibility. For example, you can create a few stores with
different products(and domains), linking to the same wallet, so no
need to input some settings for each store and syncronyze each store.
Now your wallet is starting to sync right after you create it, if you
don't leave the wallets page after creating wallet, in some time you
will see a notification in top right corner of screen saying that wallet
has been syncronyzed. If xpub was invalid, it will say about it
instantly.
Also a lot of other fixes and things have been fixed, performance was
improved.